+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/ui/screens/calendar/DayCell.kt
@@ -111,7 +111,6 @@ fun DayCell(
                 date = date,
                 addChore = {
                     addChore(it)
-                    onDismiss()
                 },
                 visible = selected && expanded,
                 viewModel = viewModel